operations and statistics moved to separate modules
